Minor fuckery with 1-2
Make it possible to deal with input files that don't have a new line
at the end. As well as add a callback function to the get_input, so
that we can return a vector of _parsed_ stuff, instead of strings,
this way we parse the strings to integers only once, instead of every
time.

Also use a HashSet as it's more appropriate