Strugglebus
Though I have an idea that has to do with Sets, and in the find best path
instead of going for all possible keys, you only go for the reachable ones
(which is what is controlled by the set), then once you get the key, if the
set has has a door opened by the appropriate key, you join the set that also
has this door. Be sure to, when joining sets, to also check if there are maybe
some others sets that need to be joined (check all the keys that you have and
all the dors in the "new set" and see if there are some unjoined sets for those doors,
repeat until no more).

Granted I need to actually fucking implement that idea. Starting with the floodfill
that generates these initial sets.

I'll probably take a short cut and assume that ALL DOORS create partitions (i.e that
there is no one door that given all other keys you can approach from both sides).
'cause dealing with that you might need to merge the sets already sooner or something.
garble garble hatefulness