day11
So first I tried some recursive accumulator nonsense, but C ain't really
made for that, and neither was this task, so I was banging my head
against the wall until I figured that I can treat this like a good old
floodfill/pathfinding sort of thing and that proved to work _much_
better. I'd almost be curious to see what others did here.