day 13
actual C programmers will kill me for not cleaning up the memory after
I'm done. That's the one part of C that's somewhat annoying though. Also
implementing a priority q as a linked list of linked lists. Which I mean
works but obviously isn't the most efficient way of doing so. granted
poping is quick (it's basically O(1) ), pushing though is basically
O(N_prio) which I guess is fine as long as the number of Ns isn't too
large. Though for the current thing it's def good enough and not a thing
that one should worry about.