Day 20
nothing difficult here, the only thing that is worth mentioning is a
note that a - b is not always what you expect when you have unsigned
integers (so umm, careful there in e.g. comparison functions and
whatnot). also MAX_NUM + 1 is again something that can overflow, so
it's better to -1 if you know that you have a large number which can
approach the MAX_NUM

Also have I mentioned how much I'm starting to like the _lack_ of magic
in C (in that you have no magical high-level types with "cool"
operations on them, and that you _have_ to implement those yourself if
you want anything of the sort).