Day 7
This was a bitch to write in C, mostly because no maps and the parsing
being annoying. Still runs in about 3ms (which granted is the slowest of
them all so far).

Also I should really free all those things I malloced instead of letting
the OS do it for me