Optimize day 7 to run under a millisecond
This pretty much boils down to "implement a hash function for the
strings". The hash function itself could probably be faster and the
magic number for the hash size could also probably be larger and thus
more generic than what we have here. Also the current implementation
does end up being a bit data specific, due to the hash size being a
magic number that was empircally derived from my data as the smallest
number for which I didn't have any colisions for my input data. It's
likely not true for generic input data.