Day 8
ah, good ol assembly (also I'm impressed that this actually runs both
things under a millisecond without me having to spend time trying to
make it faster). And that's with brute-forcing the "one change" for the
instruction for part 2. I do wonder if there's a "faster way to do
this", but honestly this feels "good enough"