Day 9
in theory part 1 could use some hashes, but in practice checking the
sums in a for loop is probably faster than implementing those hashes.

For part 2 ended up spending a lot of time debugging due to having
chosen a too small window size (500 should actually be enough, but 400
was too small). Also had a fun off-by-one issue where I wasn't checking
the last nr in the window for the min/max purposes