Day 10
part 2 is satisfying once you get it. Basically we need to look for
windows of items that can be skipped, then we just multiply the
combinations of elements in that window. Now the thing that I've
implemented works for my input where the max window size is 3 ( that is
three skippable items in a row, with the edges being -1 and +1 of the
skippable numbers. For a more genereic algo you'd have to do more
fuckery, I suppose.

Also uint64_t is required, I spent some time being dumb and not getting
why it's the wrong answer due to using unsigned int and getting an
overflow.