Day 11
runs both parts in 8ms, probably can be optimized further to have both
parts done in the same outer loop, cutting it down to like 3/4 ms, but
I'd need something more fundamental to cut it down even further.

The major optimizations are:
notice that things that don't change between two runs will stay that way
thus they can be skipped, also we can minimize the "search area", by
keeping track of the smallest and largest x and y coordinates where
changes _still_ happen.

Additionally no copying of arrays, and just swapping array pointers
around to decide which is the current and which is the previous one.

Also for part 2 the _biggest_ win is to notice that the "interesting
seats" are always the same ones, thus they can be pre-calculated and
their locations stored in an array that can be looped, thus making that
part of the code run in the same speed as the equivalent part 1 code.

In order to not have to deal with out of bounds issues, the actual grid
has "padding" of an additional row and column on each side of it. thus
as long as your for loop bounds are 1 .. <= grid_size, you're always
guaranteed that x+1, x-1  will be defined.