A second faster implementation for day 11
I shamelessly stole the idea from https://github.com/xoposhiy/aoc2020/blob/main/11_seats.cs
and adapted it for C while also adding my own optimizations (like making
it work for part2 and doing all sort of nasty tricks with pointers and
encoding the grid coordinate in the integer that is the grid item.)

This thing runs in about 3.3ms on my machine, which is a 2x improvement
over my other solution.

Gotta say the idea behind this is pretty rad, as instead of simulating
the whole thing, you just DFS the settled seats, where the dfs front is
your neighbors.