Clean up useless things from day 13 and slightly optimize it
We only need to check the buses for which we haven't found the modulus,
so instead of keeping track of that in a variable in the bus list and
checking that variable every time, we instead replace the current item
in the array with the last item in the array and decrease the array
size, thus effectively forgetting about found busses, since we no longer
care about them.

This aught to cut down on the instructions somewhat.