Simplifying and making it faster by stealing abigail code
the math checks out, and it turns out initializing less memory is
_faster_ and doing less things while inside the loop is _also_ that.

(to the surprise of noone)

The reason this works (and why I was confused). I never realized that
when saying what were the last two times the previous nr was spoken is
always current_turn - 1, _some time further in the past_ (or never). so
as long as you always set the seen value as current_turn - 1 you're
good.