Day 17
p1 is symmetrical with regards to z, but p2 is _not_, so you do need to
allocate extra memory for the negative w and z axes. but other than
that, fairly straight forward with a pair of **chars that get switched
around as the "previous" and "new" grids and filthy 4d array element
math.