Day 21
this was somehow _easy_ in comparison to a few of the previous days. One
could speed this up by having a hash and not having to do crazy
O(n) index lookups (where you basically have to compare strings n
times), but like "it's fine" and I'm too lazy to write a decent hash
function that would actually speed things up.