Day 25
turns out I had trouble reading the instructions and was trying to
square the numbers instead of always multiplying them by 7. The
hillarity was that it _works_ on the test data :D

Now of course I realized that something was wrong when I figured that
the number of iterations cannot be > 20201227 and I was doing
_billions_, thus I must be caught in a loop. Read the instructions
several more times until I _actually_ understood them.