Improve aesthetics of day4
as is custom I've tried to make the code more funcitonal and (hopefully)
shorter.

the one key thing that I've picked up that triggered this is the
`is_map_key` guard, that made checking for the bingo nicer.

also now the function that checks for bingo returns the accumulator for
the draw_numbers function, which is also made to work for both pats by
taking a mode argument which simply helps determine it's return state.