Improve the elegance of day 5
change the order of the arguments in the count_lines function, so that
we can do a cool pipelinening of the first parameter in the function
body.

Also change to a two map solution where the first map contains all the
points where a line exits at least once, and a second one where a line
exists at least twice. Having this makes incrementing the count a sexy
thing, where we do this only if a key doesn't exist in the second map,
and the checking if a key exists in a map can be done at function guard
time.

Also have the map keys (the points) be represented as simple integers,
as that increases the performance of these hash lookups (vs {x,y} tuples).