speed up day 5 by 2x by NOT using maps
So here's the deal with maps in elixir: because they are immutable,
adding a new key means you create a new map (which is an O(sizeof old_map))
with with the extra key in it. So basically if you do a lot of
Map.put(), you can expect the perf to be "not great". Contrast this with
lists, where because of immutability adding an element at the head is an
O(1) op, since all we do is create a new struct with a {value : val,
next: *old_list} and move on with our lives.

This in turn means that it's actually _much_ faster (about 2x) to simply
create a list that holds all of the points that any of the lines touch,
and then sort that list and then count the cases where the list has
items in the form of [x,x,y | rest].

Def add this to "list of things Ramūns has learned about elixir during
AOC 2021"