Ignore the nines and get rid of the seen map
so first of all the nines are useless since they can be "inferred" in
in both parts.

also for part 2 instead of having a seen map, we instead REMOVE things
from the "unvisited" map, thus we simply don't process points that are NOT
in that map, which I guess saves on memory or something and also makes
the code shorter and more elengant (e.g. I no longer need an if and can
deal with that condition in a guard)