Day 12
more of the tail recursive solutions.

Tho I have to admit the big speedup came once I realized that I was
spending a lot of time doing string comparisons and "lowecase"
conversions, and that it's more efficient to do that stuff just once
when parsing the data.

I am somewhat annoyed by how part1 and part2 has basically different
functions, but I can't think of a nice way to combine them without
making the performance worse, so I'm just gonna keep it as is