Filthy v1 of day 14
this is where I'm dumb and instead of doing smart things, I'm just using
the knowledge that my laptop can generate a list that's 20-deep, thus
I do that for the big list, and then I notice that any `AB` will always
start and end with A and B respectively and that what happens inside
them only depends on this AB, so I then generate the list for each
unseen pair, count the frequencies, cache that, and then add the
frequency map to the "total frequency count" (tho one then has to be
careful with how A and B get counted).

Anyhow, this works on an M1Max in a "reasonable" time (so seconds), but
like it's bad and wrong, and I'm only putting it here as an example of
what _can_ be done, but should absolutely _not_ be done. This will be
replaced by a saner version where we realize that we don't need to
generate the list at all :)