Elixir is insane!!!
I mean:
```
def prepare_data([input]) do
  len = byte_size(input)
  num = input |> String.to_integer(16)
  <<num::integer-size(len)-unit(4)>>
end
```

I just turned a long-ass-string into a number that I can treat like a
regular-old int IN ONE GO, like it was "FF" or something. And then, we
just IN ONE GO convert that shit to bitstream, which we then
pattern-match on.

MADNESS!!

YOU CAN'T DO THAT SHIT

I mean I _know_ where this comes from, but like once you actually try
this, it's ... fucking awesome is what it is.