make day 18 _slightly_ faster
because several hours of debugging recursion in the morning wasn't
enough, I decided that I clearly need to debug it some more, by not only
keeping track track of depth at each node, but also if it requires a
split (which makes splitting _slightly_ faster) so that we always know
where and how to do splits.

Anyhow now the data structure is {depth, needs_split?, [l, r]}, and I
guess the splitting functions are simpler now (because we don't need to
keep track if we split something or not). That said, we pay for all
this, by having to ensure that we know if any node needs a split and
propogating that up the tree.