Apply an optimization to day 24
ok, so this still satisfies _my_ constraints of "this has to work with
test data without me handling that in a special way", but it uses two
more things from the program to make the whole thing _much_ faster (it's
like two seconds now).

1) skip a few instructions (the initial `mul x 0` do nothing and no
   point of doin them)
2) and this is the _big_ one and arguably makes this _less_ generic -
   add a constraint on the z that it has to be at each level no more
   than z^level if level < 7 and z^14-level otherwise

The second one cuts the runtime from over a minute to under two seconds,
and, since it doesn't impact the test programs in any way, it's a valid
optimization under my rules :)

That said, for a "purer" implementation see the previous commit (but
enjoy waiting a lot more)

I also tried to "compile" the miniprograms into a single function,
however those turned out to actually be _much_ slower than just
executing the instructions. This is because the "compiled" version ended
up having _more_ instructions than the "just do these steps" version, so
it ended up being about an order of magnitude slower (e.g. 2000ns vs 42000ns)

Either way, the "compiler" is interesting and is included in the commit