On the last day of advent...
we moved some herds.

It turns out it's actually handy to use a similar data structure to what
we did on day 23, in terms of we have a "grid" where we keep everything,
but we also have a list (or two in this case) of stuff that we'll be
moving around.

This makes it easier in terms of "we don't have to go through the whole
grid to find the stuff that we want to move" and also we don't need to
re-create the entire grid by copying it one cell at a time, we can
actually use the previous grid as our initial state for the reducer.