Prep-work for 2022
this year we're doing perl. Tail-recursive functional perl.
The point I guess is to piss off both the FP people and the perl
people at the same time, by providing code that literally everyone
hates.

To make my life somewhat easier I wrote a script that sets up the file
for a given day (and opens it at the end) as well as downloads the input
data.

Note that the runner is already written without any explicit for loops
;)

But yes - the rules are:
NO FOR LOOPS
NO WHILE LOOPS

tail recursion via @_ mangling + goto

map, grep, sort and the like are totes kosher.

I guess try to think of variables as immutable? (tho elixir style
rebinding seems very much like mutation so 🤷‍♂️)