Day 2
ugh, modulo math at 6AM is torture, especially if you made the mistake
of doing an off-by-one modulo math.

but yeah initially for the ease of scoring for part 1 I chose to replace
things with 1,2,3 (instead of 0,1,2), and that just broke my brain when
coming up with the math for part2, I succeeded, but eventually rewrote
it with the simpler math you see here.

Otherwise it's just a simple reducer where you add to the sum based on
some rules for each iteration.