Day 5
here we spend more time parsing data, and because we manipulate the sh
out of the stack, we actually parse it twice - once for each of the
parts, so that we get to keep the same initial state.

We could of course try to clone the inital stack somewhere somehow, but
it felt like just parsing the data seemed simpler.

otherwise it's just some popping and pushing.