I really wanted to do A* so I did the first part with that
I mean over the course of fucking around I re-did my pq as a min-heap
(since that seems "cool and all"), which I mean it's nice to actually
practice implementing it since for whatever reason that's the data
structure that I don't really keep in my head, so practicing it seemed
fun and useful.

Also turns out I really forgot the important part about A* is that you
actually have to keep a track of the distances in the seen, because you
might find a shorter path to the same point after A* gets stuck behind a
cliff or something and tries to first walk around it (before going back
far enough to actually get to that same point faster). Luckily this was
actually causing issues and I realized what was happening and changed
how that part works as well.