Throw some high-school geometry and algebra at the problem
and now it's FAST. Like stupid FAST look at this debug output:
```
ramuuns@rmax 2022 % ./runner.pl 15
checking 1518415, 1128584
it's in range of sensor 21
skipped to 2403183, 2013352
checking 2403184, 2013353
it's in range of sensor 15
skipped to 2634766, 2244935
checking 483366, 2163633
it's in range of sensor 14
skipped to 1111303, 1535696
checking 1111304, 1535695
it's in range of sensor 21
skipped to 2535984, 111015
checking 483366, 2163633
it's in range of sensor 14
skipped to 1296628, 2976895
checking 1296629, 2976896
it's in range of sensor 27
skipped to 1959269, 3639536
checking 1518415, 3198682
it's in range of sensor 8
skipped to 1991725, 2725372
checking 1991726, 2725371
it's in range of sensor 12
skipped to 2510742, 2206355
checking 2510743, 2206354
it's in range of sensor 15
skipped to 2880459, 1836638
checking 2474609, 2972436
it's in range of sensor 4
skipped to 2662274, 3160101
checking 2662275, 3160102
P1: 4861076
P2: 10649103160102
Took: 0.000428915023803711
```

so what we do is basically use the fact that we're walking along a line
equasion with always an increasing X and wher the line equasion is
either y = x + C (where C == (Py-Px) or a y = -x + C (where C == (Py+Px))

and then if we're in range of some sensor we get the line equasion of
the far range (wich will be again similar) and find where do the two
lines intersect. -> skip to that x,y, repeat loop. DONE.

(also obviously skip ahead when x < 0 and stuff like that)