Cut the search space in half by not including AA in the set
OOps, that was useless and not doing so cuts the runtime in half -
HOORRAY - it's now down to like 15s from 30s (still slower than the
non-tail-recursive version, which is annoying, but oh well). Also I
could _reaaaaaaly_ use some way of caching the best paths, but somehow
haven't been able to implement it.