Got it down to 1.6 seconds
Hashtag-goodenough,

so I got the hint on the thread, about comparing all the paths, which I
initially thought was stupid, but then I realized that I AM WALKING ALL
THE PATHS ANYWAY BECAUSE OF MY DUMB ALGORITHM, I was just throwing them
away because lol.

So I decided to change my thing to keep all the paths that are ever
encountered with the (max limit = whatever), and then for each "path",
get their best version (because you can walk a -> b -> c and get a bad
score or do b -> a -> c and get a great one), and then find the max sum
where the paths don't intersect.

also because because I sort these paths by lenght I can skip ahead
whenever the sum of any two paths is less than the maximum (which is the
thing that sealed the deal).