This is gross and also incorrect for part2
granted it's only off by one on one of the things, so it might or might
not work on others input data.

I spent waaay too much time on this, and I think this is missing the
dead ends that I wen't down on (e.g. calculating a "score, which would
be an estimate on how much geodes can we even get" and things like
that).

None of my "score" ideas were able to satisfy the criteria of:
 - prunes enough darn trees
 - doesn't prune ones that actually lead to best outcomes

I am quite unhappy with myself, but oh well.

Anyway in the end I gave up a bit and saw on _the chat_ that humans were
doing the relatively dumb thing (stuff I had dismissed earlier because of unsatisfactory amount
of tree-cutting) and that it was mostly good enough (if running for
hours is good enough).